Split release tests by platform
Keep preparation limited to one-time restore/reset, move volatile checks into host, Android, and Apple runners, and share heartbeat, workload, Appium, package, and test execution infrastructure.


Copilot-Session: c25c2c8a-2652-40ae-8fcc-c232414e9eb7
